requirements.txt
# pip install -r requirements.txt

virtualenv==1.6.1
distribute==0.6.35
nose==1.3.0
pyreadline==2.0
python-dateutil==2.2
facebook-sdk==0.4.0